liveOppServationFrontend/pyproject.toml
[project]
name = "liveoppservationfrontend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jannusch Bigge",email = "jannusch.bigge@tu-dresden.de"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio (>=1.73.1,<2.0.0)",
    "grpcio-tools (>=1.73.1,<2.0.0)",
    "plotly (>=6.2.0,<7.0.0)",
    "dash (>=3.1.1,<4.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "liveoppservationfrontend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
